New function and changes to wmf-netbox plugin to support EVPN config.
This CR adds a new function to wmf-netbox.py, which is designed to
return a dictionary of interface names, and far-side loopback IPs of
devices connected on those interfaces.

The new function is used by templates I've used to connfigure the new
EVPN-based switches in the new Eqiad cage.  Specifically the goal is
to enumerate the interfaces that OSPF should be implemented on, and
provide the loopback IP of the remote devices for the EVPN iBGP.

There is no high-level attribute in Netbox to define a switch as
belonging to the EPVN fabric.  We could look at doing something like
that, the logic in this function basically looks for links that:

a) Have IP addresses configured on them
b) Has a directly connected device of type 'asw'

Additionally there are some other changes added to support the updated
templates:

1) The 'evpn' config section from YAML is exposed if it is present.
2) The VRRP VIP handling is changed to support Anycast GW (no grp id).
3) Interface description prefix logic is adjusted so that underlay
interfaces say that and not 'Core:'.

I have no doubt that there aren't better approaches to all of this,
so please consider this a 'first stab' at the config.  At least it
should demonstrate the problem at hand, I am happy to get feedback
on alternate / better approaches to make this information available
to Homer templates.

Bug: T299758
Change-Id: I35c31a860c4d8ffb7eb59fdbe57904bdd9a66dd7