Correct wmf-netbox plugin failure with patch panel front ports
The recent changes in Ie5147bcfd771408202708d097200c8a0551c8c8e have
caused homer to fail in Eqiad.  This is because some interface to
interface ('core') links there are not directly connected via a
single cable, but instead go through one or more passive patch
panels.

These are modeled in Netbox as type dcim.frontport, and
the previous change skipped those, only looking for dcim.interface
type.  The resulting None object caused an error as some of the
Jinja templates use 'startswith' which is only valid if the value
is a string.

Change-Id: I474dd9d34fb053d3a8c4eb93d723d2fe6b506d75