Interface description: don't add the z_side when disabled
I initially thought it was useful, but the issue is that the "offline"
Netbox script remove the cables in Netbox, but doesn't run Homer.
So it leaves homer with an outstanding diff, for example:
-   description "DISABLED mw2253";
+   description DISABLED;

It's not convenient to ask people to run Homer after the Offline script,
so I'm reverting this behavior to only show "DISABLED" on disabled
interfaces.

Change-Id: I73b50a30df013a27e54ea2cccde47d10036732e3