Modify wmf-netbox plugin to provide QFX5120-48Y port block speeds
This adds a new internal function to wmf-netbox.py which returns a
data structure with the speed (in Gbit/sec) of each block of 4
SFP28 ports on the device.

This platform has ports that can run at different speeds, however each
contiguous block of 4 need to be set to the same rate.  The speed for
each block of 4 is set in the 'chassis fpc' context using the ID of
the lowest port in the block (say '0' for ports 0-3).

The new function returns a dict, keyed by the lowest port number in
the block, with values representing the speed each block should be
set to in the device config / templates.

If the combination of port speeds in Netbox is invalid the function
returns None.  In combination with the config template changes in
Ic22d3ce9801e962d59e4b8d9da985a94c6e6a083 this results in a config
with no stanza for 'pic 0' being generated. So if pushed to a device
the existing settings will not be modified.

Input validation (I9f0b41912da2ba80f603a893787271739054feeb) for new
hosts, and the Network report (https://w.wiki/5nVU), should mean we
don't end up in that situation / catch it if we do anyway.

Bug: T303529
Change-Id: I33af2d4e62a7b433439ca4c1b565c8e93617f7d2