Expose additional link information to Homer templates in wmf-netbox.py
This patch modifies the existing code for creating interface
descriptions, breaking the work into two parts.

Most of the logic from the original interface_description() function
has been moved to the _get_link_data() function.  This function does
the same checks to discover the related properties of the link, but
instead of just creating a string for the description, returns a dict
of those attributes.  I've slightly refoactored  / cleaned up in the
latest patchset after digging into the logic a bit more.

Exposing the link's attributes in a dict like that means it can be
consumed elsewhere in templates.

The replacement interface_description() function builds the
description string based on the data in that dict instead.

Unlike previous patch set this one does cause some changes, due to
code optimisation suggested, just removes a comma on some
descriptions.  I'm unsure myself if it's worth it might be best as
it was:

https://phabricator.wikimedia.org/P46677

Bug: T328313
Change-Id: I1a7fc5cee3d4c3d778e891d458d04cec1a8f73fa