Return port blocks data for both QFX5120-48Y Netbox device types
We are currently modelling QFX5120-48Y devices under two separate
device types in Netbox, reflecting the old and the new Juniper SKU.

This change ensures the port block configuration is returned for
both types.

Also makes a change to ensure LAG / ae interfaces are not processed
by the port block logic, hit that error which we hadn't previously
seen.

Bug: T331519
Change-Id: I71bcb6c76835979e8657e8061900ed9139ecae45