Change 'anycast_gw' var in int config to represent type of IRB needed
Previously the 'anycast_gw' var was a boolean which was set to
indicate the interface was an IRB/Vlan int with 'anycast gw'
function.  This patch adjusts the plugin to instead make this
a string, with two values indicating which type of anycast gw
configuration is required ("single" or "vga").

The key mechanism to detect the difference is the number of IPs of
a given address fam configured on the interface in Netbox.  To
facilitate getting this number I've added a new function which
iterates over the IPs for a device and build a dict keyed by the
interface name with the IPs for each.

Bug: T350579
Change-Id: I6476975f2dab38d8b68e786612f8e70bbf14313f