Use vlan name to determine if server BGP peering should be added
This patch adds a check on the vlan name when considering if a given
server should have a BGP peering configured on a given network device.

On L3 switches it will only include the device if it is on a
rack-specific vlan.  On CRs it will only include the device if it is
connected to a row-wide one.

Check is based on our naming convention which is perhaps not ideal
but this is well controlled and it covers our case well.

Bug: T355225
Change-Id: Iebb9cd6d80492cbd24d18075f29c6553a1852e78