Skip switch interface if no untagged_vlan when finding bgp peers
I should have better tested the previous change.  Trying to access
the 'untagged_vlan' of routed ports throws an error of course.

We could just catch the error but better to express our intent -
i.e. we only want to assess ports that are L2 and have an untagged
vlan.  So check for that also.

Bug: T355225
Change-Id: I219de92958526bb626116d3f22d4defa47beae3e