Support VM BGP automation using Netbox flag for L3 POPs
As described on task the Homer automation to find BGP peers to
configure on networking devices does not work for VMs at L3 POPs.

This patch should ensure these are also covered by the automation.
It mainly does two things:

1) It does not add all VMs at a site to the list to be configured on
CRs anymore.  If a VM is at a site and has the bgp flag it will find
a ganeti host belonging to the VM's cluster and do the same check
done for physical servers - i.e. only include it if the genti host
is connected to a virtual-chassis switch or vlan using the old
row-wide naming convention.

2) When iterating over connected switch interfaces to find direct
BGP peers needed on L3 Switches it does additional steps if it
finds a host that is part of a virtualization cluster (i.e.
ganeti hosts).  In this case it will pull the list of other hosts
in the same virtualization cluser, and if they are all in the same
rack then it will add VMs belonging to the cluster with the BGP
flag set to the list of peers.

Bug: T364480
Change-Id: I47579d0f6dfe1acea2785e5053014bd05fcd5023