Add function to wmf-netbox plugin to provide QoS config data
Adds a new function _get_qos_interfaces() which iterates over
the data produced by _get_junos_interfaces() and compiles a
new dict of interfaces that need to have configuation applied
under class-of-service for network devices.

The criteria for what is added for each interface is as
described here: https://w.wiki/AVTF

The data is intended to be used in the 'interfaces' section of the
class-of-service config which can be seen in change
Ie034be7abb39a599eb0035822a72a8eb80be4d3e

This patch also adds a new var to the class, _junos_interfaces(),
which _get_junos_interfaces() now populates.  The benefit here
is that now on second or subsequent calls of the function it
can just return the existing data.

Also some small refactoring to the _get_junos_interfaces() function
to simplify the code and ensure interfaces with subs are by
default set to 'enabled' even if not present in netbox.

Bug: T339850
Change-Id: I951678748e446ff110bb97e3a0586e879da5afca