Adjust how we build list of server BGP peerings for CRs
Started looking at this due to the diff that occured when the new
ganeti hosts were added.  Seeing as the code already had a statement
to add a device if it was on a legacy vlan it seemed the simplest
to just use that for everything and simplify how things were done.

No diff on all our current devices, also makes the config generation
for CRs in for instance eqiad about 5 times quicker.

Bug: T381175
Change-Id: I41084d17b61c8172655053f303ef5eef7484d2d1