WMF-Plugin: Potential clean-up of b-end circuit finding logic
This builds on top of the change to use GraphQL to fetch the
device interfaces.  It replaces some of the old logic for finding
the b-end of devices.

Specifically we can rely on Netbox to tell us if the far-side has
a "connected_endpoint", and if that is of type "interface" (i.e.
connected to one of our own devices).  Netbox knows the far-side
we don't need to think about 'link_peers' here.

Seperately we can check the "link_peers" and identify if there is
a circuit connected to this port, and if so add the details.

Bug: T310577
Change-Id: Ia970fbfaedbadf308c2c9bc198b5a116c022c1b8