New function to generate device-specific IBGP data from cluster YAML
This patch aims to expose a simple data structure to configuration
templates which can be used to generate the IBGP and OSPF config
for L3 datacentre switches (EVPN or regular BGP).

It will use an updated model in the YAML replacing our current
'evpn_bgp' var: https://phabricator.wikimedia.org/P76660

The new function, _get_ibgp_config, parses this data and extracts
the relevant data for the current devices defined role in the
cluster.

Examples of the generated data can be seen for a variety of devices
here: https://phabricator.wikimedia.org/P76659

This is something of a WIP.  The logic is fine afaik, but it is
perhaps not ideal to re-generate 'device_data' in the plugin instead
of passing it from Homer to the plugin.

Bug: T394530
Change-Id: I1e1caf5f4815970deccb73a3ddb72b6a1cdefec3