WMF Plugin: do not process disabled ports for block speed setting
We had a slight issue with the logic here.  The wmf plugin parses
the set of ports for QFX5120 (Trident 3) devices and works out
what speed each 'block' of four should be set to in the chassis
config.

The main problem was that it was processing disabled ports -
which our input validator for the same constraint doesn't - and
failing if there were (for instance) enabled ports at 25G in the
same block as disabled ones at 10.  This patch changes to only
consider enabled ports when deriving the speed for each block.

Also added a warning log if there is an inconsistency - in which
case we return nothing to not alter the block config - so the
user will be aware if this does happen.

Bug: T394333
Change-Id: Ia2a082b25bba00bb50b3c956178af722114d908d