wmf-plugin: New function to expose generic interface data to modules
This adds a new function to the wmf-plugin called
_get_device_interfaces().  This exposes the device interface data in
a simple dict built from the returned GraphQL interfaces_list, and
then adds the additional info generated from the existing
_get_link_data(), interface_description() and interface_mtu()
functions.

It can be viewed as similar to the _get_junos_interfaces() function,
except it does a whole lot less mangling as the goal is to be vendor
neutral and it is expected any further vendor-specific mangling will
be done elsewhere.

Also adds the Nokia Leaf model we are using to the variable
L3_TOP_OF_RACK_MODELS, which I've also renamed from L3_SWITCHES_MODELS
as we are only listing the top of racks in it not spines.

Bug: T402511
Change-Id: Ie8a38b9fc4033932da08a0786050521409a3ab39