WMF-Plugin: Include the BGP role when exposing the IGBP data
Junipers automatically assume a BGP peer that does not announce a
route-reflector cluster ID when the session is established is a
route-reflector client, and reflects routes to it.

Nokia need to have peers explicitly configured as 'client' on the RR
for this to happen.  As such we have to slightly adjust the dict we
return, moving the IPs under 'addresses' and adding a new key called
'rr_client' to expose if the peer is a client or not.

Bug: T402577
Change-Id: Ie769f5392c819b7bc6545e670d843f5de5f022c9