Machine Learning beast servers: allow BGP to alternate rack
Homer was not generating the BGP config for these servers due to
this check, which only attempts to enable BGP for servers in the
same rack as the switch being configured.

The original line is there so we don't try to create BGP peers
to LVS hosts which have links to switches across the DC.  Those
will go away when low-traffic services are moved to IPIP rather
than direct-return, but for now its needed.

But the new massive ML hosts are racked as the only device in
their rack, and cabled to a switch in a separate one.  So we need
to change the conditional to include those ones.  So for now add
a check on the device model.

Change-Id: I0c20c326eb9b7dfd3a0a4551089643a99b99f8bd